website_builder/requirements.txt
requests
python-dotenv
groq